requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.4.22
charset-normalizer==3.4.7
exceptiongroup==1.3.1
execnet==2.1.2
Faker==30.8.1
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
idna==3.13
iniconfig==2.3.0
packaging==26.1
playwright==1.48.0
pluggy==1.6.0
psutil==6.1.0
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
pyee==12.0.0
Pygments==2.20.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-base-url==2.1.0
pytest-playwright==0.5.2
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
requests==2.33.1
six==1.17.0
sniffio==1.3.1
text-unidecode==1.3
tomli==2.4.1
typing_extensions==4.15.0
urllib3==2.6.3
